Fix import paths
