+++ src/api/namespaces/atomicassets/handlers/transfers.ts
@@ -1,7 +1,7 @@
-import { buildBoundaryFilter, filterQueryArgs, RequestValues } from '../../../utils';
-import { AtomicAssetsContext } from '../../index';
-import QueryBuilder from '../../../../builder';
-import { buildAssetFilter, hasAssetFilter } from '../../utils';
+import { buildBoundaryFilter, filterQueryArgs, RequestValues } from '../../utils';
+import { AtomicAssetsContext } from '../index';
+import QueryBuilder from '../../../builder';
+import { buildAssetFilter, hasAssetFilter } from '../utils';
 
 export async function getRawTransfersAction(params: RequestValues, ctx: AtomicAssetsContext): Promise<any> {
     const args = filterQueryArgs(params, {
